Mejoras en rutas en admin

+++ src/app/components/Empleados/empleados-list/empleados-list.component.ts
@@ -4,6 +4,9 @@ import { EmpleadosService } from '../../../services/Empleados/empleados.service'
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { FilterEmpleadoPipe } from '../../../pipes/empleado/filter-empleado.pipe';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmDialogsComponent } from '../../dialogs/confirm/confirm-dialogs/confirm-dialogs.component';
 
 
 
@@ -18,7 +21,12 @@ export class EmpleadosListComponent implements OnInit {
   empleados: any[] = [];
   searchText: string = '';
 
-  constructor(private empleadosService: EmpleadosService, private router: Router) {}
+  constructor(
+    private empleadosService: EmpleadosService, 
+    private router: Router,
+    private dialog: MatDialog,
+    private snackBar: MatSnackBar
+  ) {}
 
   ngOnInit(): void {
     //this.getEmpleados();
@@ -37,7 +45,18 @@ export class EmpleadosListComponent implements OnInit {
   }
 
   editEmpleado(id: number): void {
-    this.router.navigate(['/admin-panel/empleados/edit', id]);
+    const dialogRef = this.dialog.open(ConfirmDialogsComponent, {
+      data: {
+        title: 'Confirmar Edición',
+        message: `¿Está seguro de que desea editar el empleado?`,
+      },
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.router.navigate(['/admin-panel/empleados/edit', id]);
+      }
+    });
   }
 
   createEmpleado(): void {
@@ -45,16 +64,36 @@ export class EmpleadosListComponent implements OnInit {
   }
 
   deleteEmpleado(id: number): void {
-    this.empleadosService.deleteEmpleados(id).subscribe(
-      () => {
-        this.getEmpleadosWithDistritos();
+    const dialogRef = this.dialog.open(ConfirmDialogsComponent, {
+      data: {
+        title: 'Confirmar Eliminación',
+        message: `¿Está seguro de que desea eliminar el distrito?`,
       },
-      (error) => {
-        console.error('Error deleting empleado', error);
-      }
-    );
-  }
+    });
 
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.empleadosService.deleteEmpleados(id).subscribe(
+          () => {
+            this.getEmpleadosWithDistritos();
+            this.snackBar.open('Empleado eliminado con éxito', 'Cerrar', {
+              duration: 3000,
+              verticalPosition: 'bottom',
+              horizontalPosition: 'center',
+            });
+          },
+          (error) => {
+            console.error('Error deleting empleado', error);
+            this.snackBar.open('Error al eliminar el empleado', 'Cerrar', {
+              duration: 3000,
+              verticalPosition: 'bottom',
+              horizontalPosition: 'center',
+            });
+          }
+        );
+      }
+    });
+  }
 
   getEmpleadosWithDistritos(): void {
     this.empleadosService.getEmpleadosWithDistritos().subscribe(
@@ -65,5 +104,10 @@ export class EmpleadosListComponent implements OnInit {
         console.error('Error fetching empleados', error);
       }
     );
-  } 
+  }
+  
+  goToMainPanel(): void {
+    this.router.navigate(['/admin-panel'], {replaceUrl: true}); // Ajusta la ruta según tu estructura de rutas
+  }
+  
 }
